refactor: unify cascade rules R6-R10 as single source of truth
The runtime cascade in kernel_classifier/__init__.py only had R6-R8,
missing the R9/R10 escalation rules (MODERATE -> IMPORTANT) that exist
in the ML pipeline's severity_cascade.py.  This desync meant the
classifier's runtime impact predictions were weaker than training-time.

Extract the authoritative apply_cascade into a new shared module
(kernel_classifier/cascade.py) and have both the runtime classifier and
the ML pipeline import from it.  No more duplicated rule logic.

Made-with: Cursor
