+++ src/aegis_ai/kernel_classifier/__init__.py
@@ -8,7 +8,7 @@ At runtime, the classifier:
   2. Extracts 49 binary feature flags from the patches
   3. Merges 3 CVSS score features (has_cvss, cvss_score, cvss_score_bucket)
   4. Runs the XGBoost model to predict severity (IMPORTANT/MODERATE/LOW)
-  5. Applies a 3-rule severity cascade using CVSS vector components
+  5. Applies severity cascade rules R6–R10 using CVSS vector components
 
 Configuration (environment variables):
   AEGIS_KERNEL_CLASSIFIER_DIR  — path to the classifier directory containing
@@ -28,12 +28,11 @@ from typing import Optional
 import httpx
 import numpy as np
 
+from aegis_ai.kernel_classifier.cascade import SEVERITY_LABELS, apply_cascade
+
 logger = logging.getLogger(__name__)
 
 KERNEL_COMPONENTS = {"kernel", "kernel-rt"}
-
-SEVERITY_MAP = {"IMPORTANT": 0, "MODERATE": 1, "LOW": 2}
-SEVERITY_LABELS = {0: "IMPORTANT", 1: "MODERATE", 2: "LOW"}
 
 CVSS_ISSUER_PRIORITY = ["NIST", "RH", "CVEORG", "OSV", "CISA"]
 
@@ -48,57 +47,11 @@ def is_kernel_component(components: list) -> bool:
     )
 
 
-def _parse_cvss_vector(vector_str: str) -> dict[str, str]:
-    parts = {}
-    for seg in vector_str.split("/"):
-        if ":" in seg and seg != "CVSS:3.1":
-            k, v = seg.split(":", 1)
-            parts[k] = v
-    return parts
-
-
 def _score_to_bucket(score: float) -> int:
     for i, boundary in enumerate(SCORE_BUCKET_BOUNDARIES):
         if score < boundary:
             return i
     return len(SCORE_BUCKET_BOUNDARIES)
-
-
-def _apply_cascade(
-    severity: int,
-    cvss_score: float,
-    cvss_vector: str,
-    patch_flags: set[str],
-) -> int:
-    """Apply daemon-derived severity adjustment rules (R6, R7, R8).
-
-    These three rules were selected via per-rule ablation on a 292-CVE
-    held-out test set and improve accuracy without harming IMPORTANT recall.
-    """
-    if not cvss_vector or cvss_score <= 0:
-        return severity
-
-    vec = _parse_cvss_vector(cvss_vector)
-    cia_hhh = vec.get("C") == "H" and vec.get("I") == "H" and vec.get("A") == "H"
-
-    # R6: LOW → MODERATE when CVSS indicates meaningful severity
-    if severity == 2:
-        if (
-            (cvss_score >= 6.7) or (cvss_score >= 5.5 and cia_hhh)
-        ) and "bpf" not in patch_flags:
-            severity = 1
-
-    # R7: MODERATE → LOW when CVSS is very low and no kernel panic concern
-    if severity == 1 and cvss_score <= 3.9 and "kernel_panic" not in patch_flags:
-        severity = 2
-
-    # R8: MODERATE → LOW when CVSS is low-moderate with local/low-impact vector
-    if severity == 1 and cvss_score < 5.5 and vec.get("AV") == "L" and not cia_hhh:
-        c, i, a = vec.get("C", ""), vec.get("I", ""), vec.get("A", "")
-        if c in ("L", "N") and i in ("L", "N") and a in ("H", "L"):
-            severity = 2
-
-    return severity
 
 
 def _select_nist_cvss3(cvss_scores: list[dict]) -> tuple[str, float]:
@@ -300,7 +253,7 @@ class KernelImpactClassifier:
         # Build active flags set for the cascade
         active_flags = {k for k, v in patch_features.items() if v}
 
-        adjusted = _apply_cascade(raw_pred, cvss_score, cvss_vector, active_flags)
+        adjusted = apply_cascade(raw_pred, cvss_score, cvss_vector, active_flags)
 
         label = SEVERITY_LABELS.get(adjusted, "MODERATE")
         confidence = float(proba[raw_pred])
